+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/productredesign/ProductModule.kt
@@ -1,5 +1,6 @@
 package com.maksimowiczm.foodyou.feature.productredesign
 
+import com.maksimowiczm.foodyou.feature.productredesign.ui.CreateProductViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
