+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/productredesign/ui/CreateProductApp.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.productredesign
+package com.maksimowiczm.foodyou.feature.productredesign.ui
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
