+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/productredesign/ui/ProductForm.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.productredesign
+package com.maksimowiczm.foodyou.feature.productredesign.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
