+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/productredesign/ui/ProductFormEvent.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.productredesign
+package com.maksimowiczm.foodyou.feature.productredesign.ui
 
 internal sealed interface ProductFormEvent {
     data object DownloadedProductSuccessfully : ProductFormEvent
